Je sais j'avais dis que j'arrêtais
Alors j'ai essayé dans tous les sens, parfois on va faire 30 essaies qui vont passer et boom tout un coup ça marche plus impossible de savoir pourquoi mais des fois on perd 7 octets du documents qui disparraisse dans l'oblivion ...

+++ etape1/client.c
@@ -93,7 +93,7 @@ int main(int argc, char const *argv[]){
 	
 	printf("Vous envoyez le fichier %s (%d charactère(s))\n",nomFichier,tailleNomFichier );
 	printf("Ce fichier est de la taille %f MO\n",(double)tailleF/1048576);
-	if(mySend(dS,fp,tailleF,nomFichier,tailleNomFichier)==-1){
+	if(mySend(dS,fp,(int)tailleF,nomFichier,tailleNomFichier)==-1){
 		perror("mySend()");
 		fclose(fp);
 		close(dS);
